requirements.txt
streamlit
requests
Pillow
litellm
folium
streamlit-folium
shapely
geopy

